Use FUSE-native `u32` for `Readdir{,plus}Request` sizes.
GitOrigin-RevId: 58e956c58982fec29d5ea58f682af1f1502bea87

+++ examples/helloworld/helloworld.rs
@@ -159,7 +159,7 @@ where
 			return;
 		}
 
-		let mut buf = vec![0u8; request.size()];
+		let mut buf = vec![0u8; request.size() as usize];
 		let mut entries = server::ReaddirEntriesWriter::new(&mut buf);
 
 		let node_offset = NonZeroU64::new(1).unwrap();
